Upgrade iced 0.13 to 0.14
`canvas::Program::update` is rewritten to take `&iced::Event` and
return `Option<canvas::Action<Message>>` across all 12 impls (timeline,
midi editor, knob, transport, mixer controls, compose lanes/canvases).
Old `(Status, Some(msg))` tuples translated to
`Action::publish(msg).and_capture()` / `Action::capture()`.

`Space::with_width/with_height` and the two-arg `Space::new(w, h)`
constructor are gone; a parser-aware codemod folded ~330 call sites
onto the `Space::new().width(_).height(_)` builder.

`Length`/`Pixels` no longer impl `From<u16>`. `theme.rs` width
constants and local dialog widths switched to `f32`; icon `size: u16`
params get wrapped with `f32::from(_)` at the call sites that pass
them into `.size(_)`/`.width(_)`.

`Application::run_with(boot)` -> `iced::application(boot, ...).run()`
with the title moved to the `.title(_)` builder. The fixed-theme
closure traded for passing the `Theme` directly (sidesteps an iced
0.14 HRTB inference failure on the `theme(|_state| ...)` form).

`keyboard::on_key_press(|key, mods| ...)` ->
`keyboard::listen().filter_map(|event| ...)` with the
`KeyPressed { key, modifiers, .. }` arm matched explicitly.

`progress_bar::height(_)` renamed to `girth(_)`.

Verified via `cargo check --workspace`, `cargo test --workspace`,
and a smoke boot in `--demo` on each of the arrange / compose / mixer
tabs.

+++ resonance-app/src/theme.rs
@@ -275,13 +275,13 @@ pub const SECTION_BAND_HEIGHT: f32 = 22.0;
 /// collapsible area between the ruler and the regular tracks.
 pub const GLOBAL_TRACK_ROW_HEIGHT: f32 = 40.0;
 /// Track-list column width on the Arrange view.
-pub const TRACK_HEADER_WIDTH: u16 = 260;
+pub const TRACK_HEADER_WIDTH: f32 = 260.0;
 /// Standard channel strip width on the Mixer.
-pub const MIXER_STRIP_WIDTH: u16 = 132;
+pub const MIXER_STRIP_WIDTH: f32 = 132.0;
 /// Master strip width.
-pub const MASTER_STRIP_WIDTH: u16 = 156;
+pub const MASTER_STRIP_WIDTH: f32 = 156.0;
 /// Inspector column width on the Mixer.
-pub const INSPECTOR_WIDTH: u16 = 320;
+pub const INSPECTOR_WIDTH: f32 = 320.0;
 /// Right-rail column width on the Compose view.
 pub const COMPOSE_RAIL_WIDTH: u16 = 280;
